Update SingScore.py, ssGSEA.py, and minimal.py
fix: update imports of the Mapping module
SingScore.py and ssGSEA.py use the Mapping module, which has been moved from collections to collections.abc in recent python versions

fix: adjust call of tumor_deconvolve
On line 211 of minimal.py, the signature matrix is passed in as "cell_signatures" instead of "sig_matrix", leading to an error. This has been changed so that the call functions properly.